Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-position/position-absolute-in-inline-002-expected.txt

outer span container span start
container span end outer span end

FAIL absolute inside inline container location should be correct. assert_equals: left matches container expected 70.203125 but got 0
